Move zeplinUrls to common openInZeplin folder
